Separated file into two
It is now possible to use in other locations the functions that count the number of paper pixels at a row or a column of an image
